Split behaviors tests
